FLPATH-667: Add notifications-common library
