+++ plugins/notifications-backend/package.json
@@ -27,6 +27,7 @@
     "@backstage/backend-common": "0.19.1",
     "@backstage/catalog-client": "^1.4.3",
     "@backstage/config": "^1.0.8",
+    "@backstage/plugin-notifications-common": "0.1.0",
     "@types/express": "*",
     "express": "^4.17.1",
     "express-promise-router": "^4.1.0",
