+++ plugins/notifications-common/turbo.json
@@ -1,0 +1,9 @@
+{
+  "extends": ["//"],
+  "pipeline": {
+    "tsc": {
+      "outputs": ["../../dist-types/plugins/notifications-common/**"],
+      "dependsOn": ["^tsc"]
+    }
+  }
+}
