+++ plugins/notifications-frontend/package.json
@@ -1,5 +1,6 @@
 {
   "name": "@backstage/plugin-notifications-frontend",
+  "description": "Frontend part of the notifications providing the Notifications API, Notifications page, left-side menu item and other REACT components.",
   "version": "0.1.0",
   "main": "src/index.ts",
   "types": "src/index.ts",
@@ -27,6 +28,7 @@
     "@backstage/core-components": "^0.13.3",
     "@backstage/core-plugin-api": "^1.5.3",
     "@backstage/theme": "^0.4.1",
+    "@backstage/plugin-notifications-common": "0.1.0",
     "@material-ui/core": "^4.9.13",
     "@material-ui/icons": "^4.9.1",
     "@material-ui/lab": "^4.0.0-alpha.45",
